Generate JS, Documentation and thumbnail from .shape file

+++ shapes/video/color/GrayscaleAvg.js
@@ -7,7 +7,7 @@
 var video_color_GrayscaleAvg = CircuitFigure.extend({
 
    NAME: "video_color_GrayscaleAvg",
-   VERSION: "2.0.213_838",
+   VERSION: "2.0.214_839",
 
    init:function(attr, setter, getter)
    {
@@ -189,7 +189,7 @@ video_color_GrayscaleAvg = video_color_GrayscaleAvg.extend({
      **/
     onStop:function( context )
     {
-        this.worker.terminate();
+        if(this.worker) this.worker.terminate();
         delete this.worker;
         delete this.tmpContext;
         delete this.tmpCanvas;
